Added all of the classes from the UML diagram.
Might want to play around with where we implement our ENUMS.
Feel free to add your changes to master as these are our first commits/pushes
I will most likely have to end up working on what you have done because
of this so I want to be able to pull changes easily.
I will commit/add/push everything to my own branch during the day and discuss with you in the evening (so in the morning for you)
In addition to all of this, maybe it could be easier to implement directions as String constants instead of ENUMS making it easier to use them and compare them.