Added createChamber method into Level.java class
I created this chamber surrounding the 20x15 matrix with walls. This enables us to work within a chamber as the assignment requires.
In order to work within this chamber, if using the same map Array from Tile.java class we need to start Width index from 1 and Height index from 1 and set their respective sizes within loops or recursion as 18 and 13. Thus we will treat the 20x15 matrix as a 18x13 matrix while implementing the maze algorithm.